requirements.txt
# Core dependencies
python-telegram-bot==21.2
openai==1.23.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
aiohttp==3.9.3
schedule==1.2.1
watchdog==4.0.0  # For module hot-reloading
loguru==0.7.2   # Advanced logging with rotation

# For development and testing
pytest==8.1.1
black==24.3.0
flake8==7.0.0
